feat: mainPage에만 스크롤 스타일 적용
